Refactors the harvest commands outside of the harvest directory to resolve paths
