+++ fabfile.py
@@ -4,7 +4,7 @@ from fabric import task
 
 from environments.surfpol import create_configuration_and_session
 from postgres.tasks_remote import restore_snapshot, setup_postgres, create_snapshot
-from harvester.tasks_remote import create_super_user, connect_uwsgi, connect_flower
+from commands.projects.harvester.fabric import create_super_user, connect_uwsgi, connect_flower
 from legacy import download_database
