+++ harvester/harvester/settings.py
@@ -334,6 +334,14 @@ CELERY_BEAT_SCHEDULE = {
         ),
         'args': ("epsilon", environment.harvester.import_dataset.role)
     },
+    'harvest': {  # TODO: when refactoring the configuration we should give harvest its own time slot
+        'task': 'harvest',
+        'schedule': crontab(
+            hour=environment.harvester.import_dataset.hour,
+            minute=environment.harvester.import_dataset.minute,
+        ),
+        'args': tuple()
+    },
 }
